+++ src/scenes/LevelCompleteScene.ts
@@ -1,4 +1,5 @@
 import Phaser from 'phaser'
+import { enterMenuMusic, addMusicControl } from '../audio/music'
 import { GAME_WIDTH, SCENE } from '../constants'
 import { levels, nextLevel } from '../data/levels'
 import { addButton } from '../ui/button'
@@ -24,6 +25,8 @@ export class LevelCompleteScene extends Phaser.Scene {
   }
 
   create(data: CompleteData): void {
+    enterMenuMusic(this)
+    addMusicControl(this)
     this.add.rectangle(GAME_WIDTH / 2, 270, 960, 540, 0x1b1e2b, 0.55)
     this.add.rectangle(GAME_WIDTH / 2, 289, 480, 378, 0x252542).setStrokeStyle(2, 0x92c8ee)